Fixed issue with db requests. App is in a working state for data entry.
TODO1: Add logic that marks dates based on the recorded dates within the
daily_expenses table. This allows dates to be marked if the application
is brought down and brought up again.

TODO2: Start working on the kubernetes sides of things, or continue to
develop the app to show results of data entry. For the time being, data
can be entered daily.